linker: Allow devices to force shim libs
There are certain contexts in which the environment is cleansed.
Two examples that I know of are gps processes and mali gles blobs.

Generally, it is a better idea to use the environment variable because
then you can customize it on a per-service / script level and it is
easier to test and debug changes on the fly.

However, to avoid having libdimytry hexedit these blobs, allow a
last resort device level forced list of shim libraries.

Change-Id: I2f6aff9325beb5aa2f748bf72e6c3c0535d5aac2
Signed-off-by: Eliminater74 <eliminater74@gmail.com>